Move priors into separate files
